+++ tests/RateLimiterTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace RateLimit\Tests;
 
 use PHPUnit\Framework\TestCase;
-use RateLimit\QuotaPolicy;
+use RateLimit\Rate;
 use RateLimit\RateLimiter;
 
 abstract class RateLimiterTest extends TestCase
@@ -17,7 +17,7 @@ abstract class RateLimiterTest extends TestCase
      */
     public function it_maintains_number_of_remaining_attempts(): void
     {
-        $status = $this->getRateLimiter()->handle('test', QuotaPolicy::perMinute(10));
+        $status = $this->getRateLimiter()->handle('test', Rate::perMinute(10));
 
         $this->assertFalse($status->quotaExceeded());
         $this->assertEquals(9, $status->getRemainingAttempts());
@@ -29,11 +29,11 @@ abstract class RateLimiterTest extends TestCase
     public function it_returns_correct_status_when_quota_is_exceeded(): void
     {
         $rateLimiter = $this->getRateLimiter();
-        $quotaPolicy = QuotaPolicy::perHour(1);
+        $rate = Rate::perHour(1);
 
-        $rateLimiter->handle('test', $quotaPolicy);
-        $rateLimiter->handle('test', $quotaPolicy);
-        $status = $rateLimiter->handle('test', $quotaPolicy);
+        $rateLimiter->handle('test', $rate);
+        $rateLimiter->handle('test', $rate);
+        $status = $rateLimiter->handle('test', $rate);
 
         $this->assertTrue($status->quotaExceeded());
         $this->assertSame(2, $status->getCurrent());
@@ -43,16 +43,16 @@ abstract class RateLimiterTest extends TestCase
     /**
      * @test
      */
-    public function it_resets_quota_after_interval_defined_by_policy(): void
+    public function it_resets_quota_after_rate_interval(): void
     {
         $rateLimiter = $this->getRateLimiter();
-        $quotaPolicy = QuotaPolicy::perSecond(10);
+        $rate = Rate::perSecond(10);
 
-        $rateLimiter->handle('test', $quotaPolicy);
-        $rateLimiter->handle('test', $quotaPolicy);
-        $rateLimiter->handle('test', $quotaPolicy);
+        $rateLimiter->handle('test', $rate);
+        $rateLimiter->handle('test', $rate);
+        $rateLimiter->handle('test', $rate);
         sleep(2);
-        $status = $rateLimiter->handle('test', $quotaPolicy);
+        $status = $rateLimiter->handle('test', $rate);
 
         $this->assertFalse($status->quotaExceeded());
         $this->assertSame(9, $status->getRemainingAttempts());