+++ tests/TestAsset/InMemoryRateLimiter.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace RateLimit\Tests\TestAsset;
 
 use DateTimeImmutable;
-use RateLimit\QuotaPolicy;
+use RateLimit\Rate;
 use RateLimit\RateLimiter;
 use RateLimit\Status;
 
@@ -14,23 +14,23 @@ final class InMemoryRateLimiter implements RateLimiter
     /** @var array */
     private $store = [];
 
-    public function handle(string $identifier, QuotaPolicy $quotaPolicy): Status
+    public function handle(string $identifier, Rate $rate): Status
     {
-        $key = "$identifier:{$quotaPolicy->getInterval()}";
+        $key = "$identifier:{$rate->getInterval()}";
 
         if (!isset($this->store[$key]) || time() > $this->store[$key]['expires']) {
             $this->store[$key] = [
                 'current' => 1,
-                'expires' => time() + $quotaPolicy->getInterval(),
+                'expires' => time() + $rate->getInterval(),
             ];
-        } elseif ($this->store[$key]['current'] <= $quotaPolicy->getQuota()) {
+        } elseif ($this->store[$key]['current'] <= $rate->getQuota()) {
             $this->store[$key]['current']++;
         }
 
         return Status::from(
             $identifier,
             $this->store[$key]['current'],
-            $quotaPolicy,
+            $rate,
             new DateTimeImmutable('@' . $this->store[$key]['expires'])
         );
     }